requirements-dev.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]
uvicorn[standard]
gunicorn

asyncpg

sqlalchemy
sqlalchemy_utils

pydantic==1.8.2
strawberry-graphql
click==8.1.3
aiodataloader

https://github.com/hrbolek/uoishelpers/archive/refs/heads/main.zip

pytest
pytest-cov
pytest-asyncio
aiosqlite